Allow slash in volume names
In the past, I have had problems with restores failing due to the device
for differential volumes can't find the volumes stored by the incremental.

I also have not been quite satisfied with having all the files in one
large directory without a structure, e.g., when I have different
retention policy for some volumes.

My fix is quite simple: add '/' to the list of acceptable characters in
a volume name.  this allows me to specify the same Archive Device for
all the file media devices, and do the separation with Label Format like
"incr/Incr-". Restore is unproblematic in this setup, since any Device
can be used to access all the volumes (even if they are actually in
different filesystems, due to the structure of my mounts).

Signed-off-by: Marco van Wieringen <mvw@planets.elm.net>